+++ drivers/hardtimer.h
@@ -2,7 +2,7 @@
 #define __HARDTIMER_H__
 #include <stm32f4xx.h>
 void Timer4_init(void);
-double Timer4_GetSec(void);
+float Timer4_GetSec(void);
 void timer_init(TIM_TypeDef * tim);
 void delay_us(TIM_TypeDef * tim,u16 us);
 #endif
